requirements.txt
aiogram==3.15.0
python-dotenv==1.1.0
openpyxl==3.1.5
